requirements.txt
bs4
requests
pydantic
pyaml
pytest